Add files via upload
This code is for generating watermarked OFDM signals. We use pulse amplitude modulation to embed the pseudonyms on the top of the OFDM data signals. 
Then we use two SDR nodes running GNU Radio with an RF link. On one of the nodes, we transmit the watermarked OFDM signal and on the other node, we collect IQ samples. Then the collected OFDM samples are post-processed to detect the pseudonym bits.